# internal/httpapi/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Discovery Service API
  description: |
    Standalone orchestrator discovery service. Aggregates subgraph, ClickHouse,
    discover API, and optional sources into a materialized Postgres dataset.
    Clients send filters, sort, and topN at query time — no server-side discovery plans.
  version: 1.0.0
  contact:
    name: Livepeer Discovery Service
servers:
  - url: http://localhost:8088
    description: Local development
  - url: https://discovery.example.com
    description: Production (replace with your public URL)

tags:
  - name: Health
  - name: Discovery
  - name: Dataset

paths:
  /healthz:
    get:
      tags: [Health]
      summary: Liveness probe
      operationId: healthz
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/discovery/health:
    get:
      tags: [Health]
      summary: Discovery health alias
      operationId: discoveryHealth
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/discovery/freshness:
    get:
      tags: [Discovery]
      summary: Materialized dataset freshness
      operationId: getFreshness
      responses:
        "200":
          description: Dataset stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreshnessResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  /v1/discovery/capabilities:
    get:
      tags: [Discovery]
      summary: List capabilities in the materialized dataset
      operationId: listCapabilities
      parameters:
        - name: serviceType
          in: query
          description: Filter by service class (`legacy`, `registry`). Omit for both.
          schema:
            type: string
            enum: [legacy, registry]
      responses:
        "200":
          description: Capability names
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  /v1/discovery/query:
    post:
      tags: [Discovery]
      summary: Query orchestrators with client-driven filters
      description: |
        Filter, rank, and return orchestrators per capability from the materialized dataset.
        Plan definitions live in the client; pass filters, sortBy, slaWeights, and topN per request.
      operationId: discoveryQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: Ranked results per capability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /v1/discovery/raw:
    get:
      tags: [Discovery]
      summary: Webhook-compatible orchestrator list
      description: |
        Returns `{address, score, capabilities}[]` for go-livepeer gateway `-orchWebhookUrl` compatibility.
      operationId: discoveryRaw
      parameters:
        - name: caps
          in: query
          description: Capability filter (repeat for multiple)
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: capability
          in: query
          description: Alias for caps
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: serviceType
          in: query
          description: Filter by service class (`legacy`, `registry`). Omit for both.
          schema:
            type: string
            enum: [legacy, registry]
      responses:
        "200":
          description: Webhook-shaped orchestrator list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookOrchestrator"
        "500":
          $ref: "#/components/responses/InternalError"

  /v1/discovery/dataset/refresh:
    post:
      tags: [Dataset]
      summary: Refresh materialized dataset from upstream sources
      description: |
        Fetches all enabled sources in parallel, resolves conflicts, and replaces the Postgres dataset.
        Requires `Authorization: Bearer <CRON_SECRET>` or `X-Cron-Secret` header when CRON_SECRET is configured.
      operationId: refreshDataset
      security:
        - CronBearer: []
        - CronHeader: []
      parameters:
        - name: X-Refreshed-By
          in: header
          schema:
            type: string
          description: Audit label for who triggered the refresh
      responses:
        "200":
          description: Refresh completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    CronBearer:
      type: http
      scheme: bearer
      description: Set to CRON_SECRET value
    CronHeader:
      type: apiKey
      in: header
      name: X-Cron-Secret
      description: Alternative to Bearer auth for cron jobs

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid cron secret
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok

    FreshnessResponse:
      type: object
      properties:
        populated:
          type: boolean
        refreshedAt:
          type: integer
          format: int64
          description: Unix epoch milliseconds
        refreshedBy:
          type: string
        ageMs:
          type: integer
          format: int64
        totalRows:
          type: integer
        capabilityCount:
          type: integer

    CapabilitiesResponse:
      type: object
      properties:
        capabilities:
          type: array
          items:
            type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CapabilityEntry"

    CapabilityEntry:
      type: object
      properties:
        serviceType:
          type: string
          enum: [legacy, registry]
        capability:
          type: string
        offeringIds:
          type: array
          items:
            type: string

    Filters:
      type: object
      properties:
        gpuRamGbMin:
          type: number
          format: double
        gpuRamGbMax:
          type: number
          format: double
        priceMax:
          type: number
          format: double
        maxAvgLatencyMs:
          type: number
          format: double
        maxSwapRatio:
          type: number
          format: double
          minimum: 0
          maximum: 1

    SLAWeights:
      type: object
      properties:
        latency:
          type: number
          format: double
          minimum: 0
          maximum: 1
        swapRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
        price:
          type: number
          format: double
          minimum: 0
          maximum: 1

    QueryRequest:
      type: object
      required: [capabilities]
      properties:
        capabilities:
          type: array
          items:
            type: string
          minItems: 1
          example: ["streamdiffusion-sdxl"]
        serviceTypes:
          type: array
          items:
            type: string
            enum: [legacy, registry]
          description: Filter dataset rows by service class. Omit for both.
        topN:
          type: integer
          minimum: 1
          maximum: 1000
          default: 10
        filters:
          $ref: "#/components/schemas/Filters"
        slaWeights:
          $ref: "#/components/schemas/SLAWeights"
        slaMinScore:
          type: number
          format: double
        sortBy:
          type: string
          enum: [slaScore, latency, price, swapRate, avail]
          default: slaScore

    DatasetRow:
      type: object
      properties:
        serviceType:
          type: string
          enum: [legacy, registry]
        ethAddress:
          type: string
        offeringId:
          type: string
        interactionMode:
          type: string
        workUnit:
          type: string
        pricePerUnitWei:
          type: string
        orchUri:
          type: string
        gpuName:
          type: string
        gpuGb:
          type: number
        avail:
          type: number
        totalCap:
          type: number
        pricePerUnit:
          type: number
        bestLatMs:
          type: number
          nullable: true
        avgLatMs:
          type: number
          nullable: true
        swapRatio:
          type: number
          nullable: true
        avgAvail:
          type: number
          nullable: true
        score:
          type: number
        slaScore:
          type: number
          nullable: true

    QueryResponse:
      type: object
      properties:
        results:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/DatasetRow"
        datasetVersion:
          type: integer
          format: int64
        queryTimeMs:
          type: integer
          format: int64
        sourceFreshness:
          $ref: "#/components/schemas/SourceFreshness"

    SourceFreshness:
      type: object
      properties:
        refreshedAt:
          type: integer
          format: int64
        refreshedBy:
          type: string
        ageMs:
          type: integer
          format: int64
        capabilityCount:
          type: integer

    WebhookOrchestrator:
      type: object
      properties:
        address:
          type: string
          description: Orchestrator service URL
        score:
          type: number
          format: float
        capabilities:
          type: array
          items:
            type: string

    RefreshResult:
      type: object
      properties:
        refreshed:
          type: boolean
        capabilities:
          type: integer
        orchestrators:
          type: integer
        durationMs:
          type: integer
          format: int64